# back-end/cmd/web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "wowcollector.io API description",
        "title": "wowcollector.io",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "paths": {
        "/api/v1/battle-net/realms-regions": {
            "get": {
                "description": "Get all the realms and regions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BattleNet"
                ],
                "summary": "Fetch all realms ang regions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RegionRealmResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/battle-net/root-achievement-categories": {
            "get": {
                "description": "Get all the achievement categories which are root (top level)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BattleNet"
                ],
                "summary": "Fetch all root achievement categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.AchievementCategoryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/character/{region}/{realm}/{character}": {
            "get": {
                "description": "Get summary information about a character",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Character"
                ],
                "summary": "Fetch character profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Region",
                        "name": "region",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Realm",
                        "name": "realm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Character",
                        "name": "character",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CharacterProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/character/{region}/{realm}/{character}/achievements": {
            "get": {
                "description": "Get achievement collection for character",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Character"
                ],
                "summary": "Fetch character achievement collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Region",
                        "name": "region",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Realm",
                        "name": "realm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Character",
                        "name": "character",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Root category to include",
                        "name": "rootCategoryId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.AchievementCollectionCategorySwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/character/{region}/{realm}/{character}/mounts": {
            "get": {
                "description": "Get mount collection for character",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Character"
                ],
                "summary": "Fetch character mount collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Region",
                        "name": "region",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Realm",
                        "name": "realm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Character",
                        "name": "character",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.MountCollectionCategorySwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/character/{region}/{realm}/{character}/toys": {
            "get": {
                "description": "Get toy collection for character",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Character"
                ],
                "summary": "Fetch character toy collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Region",
                        "name": "region",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Realm",
                        "name": "realm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Character",
                        "name": "character",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.ToyCollectionCategorySwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/leaderboard/mounts": {
            "get": {
                "description": "Get the leaderboard mounts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BattleNet"
                ],
                "summary": "Fetch leaderboard for mounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PaginationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorresponse.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "blizzarddata.BattleNetRegion": {
            "type": "string",
            "enum": [
                "eu",
                "us"
            ],
            "x-enum-varnames": [
                "REGION_EU",
                "REGION_US"
            ]
        },
        "errorcodes.ErrorCode": {
            "type": "string",
            "enum": [
                "CHARACTER_NOT_FOUND",
                "LOADING_BATTLE_NET_DATA",
                "INTERNAL_ERROR",
                "NO_DEFAULT_MOUNT_VIEW",
                "NO_DEFAULT_TOY_VIEW"
            ],
            "x-enum-varnames": [
                "CHARACTER_NOT_FOUND",
                "LOADING_BATTLE_NET_DATA",
                "INTERNAL_ERROR",
                "NO_DEFAULT_MOUNT_VIEW",
                "NO_DEFAULT_TOY_VIEW"
            ]
        },
        "errorresponse.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/errorcodes.ErrorCode"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.AchievementCategoryResponse": {
            "type": "object",
            "properties": {
                "displayOrder": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.AchievementCollectionAchievement": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "displayOrder": {
                    "type": "integer"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isCompleted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                }
            }
        },
        "response.AchievementCollectionCategorySwagger": {
            "type": "object",
            "properties": {
                "achievements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.AchievementCollectionAchievement"
                    }
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "displayOrder": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.CharacterProfileResponse": {
            "type": "object",
            "properties": {
                "faction": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "realm": {
                    "type": "string"
                }
            }
        },
        "response.LeaderboardEntry": {
            "type": "object",
            "properties": {
                "character": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "realm": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                }
            }
        },
        "response.MountCollectionCategorySwagger": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "mounts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MountCollectionMount"
                    }
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                }
            }
        },
        "response.MountCollectionMount": {
            "type": "object",
            "properties": {
                "assets": {
                    "$ref": "#/definitions/response.MountCollectionMountAssets"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isCollected": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.MountCollectionMountAssets": {
            "type": "object",
            "properties": {
                "display": {
                    "type": "string"
                },
                "largeIcon": {
                    "type": "string"
                },
                "smallIcon": {
                    "type": "string"
                }
            }
        },
        "response.PaginationResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.LeaderboardEntry"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.RealmResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "region": {
                    "$ref": "#/definitions/blizzarddata.BattleNetRegion"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "response.RegionRealmResponse": {
            "type": "object",
            "properties": {
                "realms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.RealmResponse"
                    }
                },
                "regions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.RegionResponse"
                    }
                }
            }
        },
        "response.RegionResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "$ref": "#/definitions/blizzarddata.BattleNetRegion"
                }
            }
        },
        "response.ToyCollectionCategorySwagger": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "toys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ToyCollectionToy"
                    }
                }
            }
        },
        "response.ToyCollectionToy": {
            "type": "object",
            "properties": {
                "assets": {
                    "$ref": "#/definitions/response.ToyCollectionToyAssets"
                },
                "id": {
                    "type": "integer"
                },
                "isCollected": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.ToyCollectionToyAssets": {
            "type": "object",
            "properties": {
                "largeIcon": {
                    "type": "string"
                }
            }
        }
    }
}